refactor: move faqs
